Adcionado arquivos de cargas e novas entidades.

+++ src/main/java/com/chicken/repository/EnderecosRepository.java
@@ -3,5 +3,5 @@ package com.chicken.repository;
 import com.chicken.entity.Enderecos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecosRepository extends JpaRepository<Enderecos, Integer> {
+public interface EnderecosRepository extends JpaRepository<Enderecos, Long> {
 }
