+++ src/main/java/com/chicken/repository/EquipamentosRepository.java
@@ -3,5 +3,5 @@ package com.chicken.repository;
 import com.chicken.entity.Equipametos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipamentosRepository extends JpaRepository <Equipametos, Integer> {
+public interface EquipamentosRepository extends JpaRepository <Equipametos, Long> {
 }
